+++ po_projects/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 """Django application to manage PO files for project translations"""
-__version__ = '0.7.0'
+__version__ = '0.7.1'