+++ po_projects/forms/crispies.py
@@ -1,12 +1,18 @@
+# -*- coding: utf-8 -*-
 """
 Crispy forms layouts
 """
+import re
+
 from django import forms
 from django.utils.translation import ugettext as _
 
+from django.utils.text import normalize_newlines
+from django.utils.html import escape
+
 from crispy_forms.helper import FormHelper
 from crispy_forms.bootstrap import UneditableField
-from crispy_forms_foundation.layout import Layout, Row, Column, HTML, Div, Field, ButtonHolder, ButtonHolderPanel, ButtonGroup, Panel, Submit
+from crispy_forms_foundation.layout import TEMPLATE_PACK, Layout, Row, Column, HTML, Div, Field, ButtonHolder, ButtonHolderPanel, ButtonGroup, Panel, Submit
 
 def SimpleRowColumn(field, *args, **kwargs):
     """
@@ -20,10 +26,32 @@ def SimpleRowColumn(field, *args, **kwargs):
 
 class SourceTextField(UneditableField):
     """
-    Layout object for rendering template field as simple html text
+    Layout object for rendering template field value as html and a hidden input
+    
+    Field value rendering is at your charge in the template (default is to use 
+    Django's 'linebreaks' filter)
     """
     template = "po_projects/input_as_text.html"
+    
+    def linebreaks(self, value, autoescape=True):
+        """
+        Converts newlines into <p> and <br />s. Additionnaly add a carriage return 
+        character before each simple line break.
+        
+        Taken from ``django.utils.html.linebreaks`` then modified to fit needs.
+        """
+        value = normalize_newlines(value)
+        paras = re.split('\n{2,}', value)
+        if autoescape:
+            paras = ['<p>%s</p>' % escape(p).replace('\n', u'<span class="explicit_cr">↵</span><br />') for p in paras]
+        else:
+            paras = ['<p>%s</p>' % p.replace('\n', u'<span class="explicit_cr">↵</span><br />') for p in paras]
+        return '\n\n'.join(paras)
 
+    def render(self, form, form_style, context, template_pack=TEMPLATE_PACK):
+        context['source_content'] = self.linebreaks(form.instance.template.message)
+
+        return super(SourceTextField, self).render(form, form_style, context, template_pack=TEMPLATE_PACK)
 
 def translation_helper(instance=None, prefix='', form_tag=False):
     """